+++ packages/drivers/local-driver/src/localService.ts
@@ -14,15 +14,17 @@ import {
 import { ContainerRuntime } from "@fluidframework/container-runtime/internal";
 import type { IRequest } from "@fluidframework/core-interfaces";
 import { assert } from "@fluidframework/core-utils/internal";
+import type {
+	DataStoreKind,
+	DataStoreRegistry,
+	FluidContainerAttached,
+	FluidContainerWithService,
+	Registry,
+	ServiceClient,
+} from "@fluidframework/driver-definitions/internal";
 import {
 	type ContainerRuntimeLoader,
 	type ContainerRuntimeLoaderParams,
-	type DataStoreKind,
-	type DataStoreRegistry,
-	type FluidContainerAttached,
-	type FluidContainerWithService,
-	type Registry,
-	type ServiceClient,
 	type ServiceOptions,
 	makeCodeLoader,
 	makeServiceClientImpl,
