+++ packages/runtime/runtime-definitions/api-report/runtime-definitions.legacy.alpha.api.md
@@ -22,9 +22,6 @@ export type AttributionKey = OpAttributionKey | DetachedAttributionKey | LocalAt
 // @alpha @sealed
 export type Audience = IAudience;
 
-// @alpha
-export function basicKey<T>(type: string): RegistryKey<T, T>;
-
 // @alpha @sealed @legacy
 export interface ContainerRuntimeBaseAlpha extends IContainerRuntimeBase {
 }
@@ -54,43 +51,11 @@ export enum CreateSummarizerNodeSource {
     Local = 2
 }
 
-// @alpha @sealed
-export interface DataStoreCreator {
-    createDataStore<T>(kind: DataStoreKey<T>): Promise<T>;
-}
-
-// @alpha @input
-export type DataStoreKey<T, TAll = unknown> = RegistryKey<Promise<DataStoreKind<T>>, Promise<DataStoreKind<TAll>>>;
-
-// @alpha @sealed
-export interface DataStoreKind<out T = unknown> extends DataStoreKey<T>, ErasedBaseType<readonly ["DataStoreKind", T]> {
-}
-
-// @alpha @input
-export type DataStoreRegistry<out T = unknown> = Registry<Promise<DataStoreKind<T>>>;
-
 // @beta @legacy
 export interface DetachedAttributionKey {
     id: 0;
     // (undocumented)
     type: "detached";
-}
-
-// @alpha @sealed
-export interface FluidContainer<TData = unknown> extends DataStoreCreator, ErasedBaseType<readonly ["FluidContainer", TData]> {
-    close(): void;
-    readonly data: TData;
-    readonly id?: string | undefined;
-}
-
-// @alpha @sealed
-export interface FluidContainerAttached<T = unknown> extends FluidContainer<T> {
-    readonly id: string;
-}
-
-// @alpha @sealed
-export interface FluidContainerWithService<T = unknown> extends FluidContainer<T> {
-    attach(): Promise<FluidContainerAttached<T>>;
 }
 
 // @beta @legacy
@@ -462,26 +427,6 @@ export interface OpAttributionKey {
 
 // @beta @legacy
 export type PackagePath = readonly string[];
-
-// @alpha @input
-export type Registry<T> = (type: string) => T;
-
-// @alpha @sealed @input
-export interface RegistryKey<TOut, TIn = unknown> {
-    adapt(value: TIn): TOut;
-    readonly type: string;
-}
-
-// @alpha
-export function registryLookup<TOut, TIn>(registry: Registry<TIn>, key: RegistryKey<TOut, TIn>): TOut;
-
-// @alpha @sealed
-export interface ServiceClient {
-    createContainer<T>(root: DataStoreKind<T>): Promise<FluidContainerWithService<T>>;
-    // (undocumented)
-    createContainer<T>(root: DataStoreKey<T>, registry: DataStoreRegistry): Promise<FluidContainerWithService<T>>;
-    loadContainer<T>(id: string, root: DataStoreKind<T> | DataStoreRegistry<T>): Promise<FluidContainerAttached<T>>;
-}
 
 // @alpha @input
 export interface ServiceOptions {
