+++ packages/service-clients/azure-client/src/azureService.ts
@@ -15,17 +15,19 @@ import type {
 	ITelemetryBaseLogger,
 } from "@fluidframework/core-interfaces";
 import { assert } from "@fluidframework/core-utils/internal";
+import type {
+	DataStoreKind,
+	DataStoreRegistry,
+	FluidContainerAttached,
+	FluidContainerWithService,
+	Registry,
+	ServiceClient,
+} from "@fluidframework/driver-definitions/internal";
 import { RouterliciousDocumentServiceFactory } from "@fluidframework/routerlicious-driver/internal";
 import {
 	type ContainerRuntimeLoader,
 	type ContainerRuntimeLoaderParams,
-	type DataStoreKind,
-	type DataStoreRegistry,
-	type FluidContainerAttached,
-	type FluidContainerWithService,
 	type MinimumVersionForCollab,
-	type Registry,
-	type ServiceClient,
 	makeCodeLoader,
 	makeServiceClientImpl,
 	rootDataStoreId,
