+++ packages/tools/devtools/devtools/src/index.ts
@@ -29,9 +29,9 @@ import {
 	type IDevtoolsLogger,
 	initializeDevtools as initializeDevtoolsBase,
 } from "@fluidframework/devtools-core/internal";
+import type { FluidContainer } from "@fluidframework/driver-definitions/internal";
 import type { IFluidContainer } from "@fluidframework/fluid-static";
 import { isInternalFluidContainer } from "@fluidframework/fluid-static/internal";
-import type { FluidContainer } from "@fluidframework/runtime-definitions/internal";
 import { ServiceContainerBase } from "@fluidframework/runtime-definitions/internal";
 
 /**
